Farukon_2::data_engine

Module ohlcv_soa_generated

Structs

OHLCVListSOA

OHLCVListSOAArgs

OHLCVListSOABuilder

OHLCVSOA

OHLCVSOAArgs

OHLCVSOABuilder

Enums

OHLCVListSOAOffset

OHLCVSOAOffset

Constants

OHLCVLIST_SOA_EXTENSION

Functions

finish_ohlcvlist_soa_buffer

finish_size_prefixed_ohlcvlist_soa_buffer

root_as_ohlcvlist_soa

Verifies that a buffer of bytes contains a OHLCVListSOA and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_ohlcvlist_soa_unchecked.

root_as_ohlcvlist_soa_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a OHLCVListSOA and returns it.

root_as_ohlcvlist_soa_with_opts

Verifies, with the given options, that a buffer of bytes contains a OHLCVListSOA and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_ohlcvlist_soa_unchecked.

size_prefixed_root_as_ohlcvlist_soa

Verifies that a buffer of bytes contains a size prefixed OHLCVListSOA and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_ohlcvlist_soa_unchecked.

size_prefixed_root_as_ohlcvlist_soa_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a size prefixed OHLCVListSOA and returns it.

size_prefixed_root_as_ohlcvlist_soa_with_opts

Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed OHLCVListSOA and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_ohlcvlist_soa_unchecked.