Farukon_2::data_engine::ohlcv_soa_generated

Function size_prefixed_root_as_ohlcvlist_soa_unchecked

pub unsafe fn size_prefixed_root_as_ohlcvlist_soa_unchecked(
    buf: &[u8],
) -> OHLCVListSOA<'_>

Assumes, without verification, that a buffer of bytes contains a size prefixed OHLCVListSOA and returns it.

Safety

Callers must trust the given bytes do indeed contain a valid size prefixed OHLCVListSOA.